한글 폰트 추가, 나레이션 텍스트 추가, 아이템 시스템 추가

+++ Assets/Scripts/WaveManage/EnemySpawnRatio.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 public class EnemySpawnRatio
 {
     public EnemyType enemyType; // 적의 타입 (슬라임, 문어, 코끼리 등)
-    [Range (0f,1f)]public int spawnRatio; // 해당 적이 스폰될 확률 (0~100)
+    [Range (0f,1f)]public float spawnRatio; // 해당 적이 스폰될 확률 (0~100)
     
     public EnemySpawnRatio(EnemyType type, int ratio)
     {
